+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+- Analyze Callers and Callees of subroutines
 
 ## [v0.1.5]
 - Added new Assignment Operators: &=, |=, >>=, <<=